requirements.txt
mss
opencv-python
pillow
pytesseract
numpy==1.21.0
pandas==1.3.0
scikit-learn==0.24.2
tensorflow==2.6.0
poker-eval==0.2.0
python-poker-engine==0.3.0
joblib==1.0.1
